+++ src/main/java/com/teltech/mathapp/operation/DivideOperation.java
@@ -2,7 +2,7 @@ package com.teltech.mathapp.operation;
 
 public class DivideOperation implements MathOperation {
     @Override
-    public Integer calculate(Integer x, Integer y) {
+    public Double calculate(Double x, Double y) {
         if (y == 0) {
             throw new ArithmeticException("Division by zero is not possible");
         }
