+++ src/main/java/com/teltech/mathapp/operation/MathOperation.java
@@ -1,5 +1,5 @@
 package com.teltech.mathapp.operation;
 
 public interface MathOperation {
-    Integer calculate(Integer x, Integer y);
+    Double calculate(Double x, Double y);
 }
