+++ src/main/java/com/teltech/mathapp/service/MathService.java
@@ -3,6 +3,6 @@ package com.teltech.mathapp.service;
 import com.teltech.mathapp.model.MathResponse;
 
 public interface MathService {
-    MathResponse calculate(String operation, Integer x, Integer y);
+    MathResponse calculate(String operation, Double x, Double y);
 
 }
